Add: Image component.Change: static pictures to use cloudinary for optimization.
